profile: Install libssl-dev on linux.
Required to avoid compilation errors on Linux machines after
https://vanadium-review.googlesource.com/#/c/14711/ is submitted.

I'm stuffing the openssl dependency into the "syncbase" profile for now.
This isn't clean since "openssl" is not a syncbase-specific dependency.
However, at this time "syncbase" is a 'required' profile and there
is no other profile that is required.

I believe there are plans to refactor this profile installation
business. I'll follow up with the folks doing that to see if there
is a more appropriate location for this dependency.

Change-Id: Iab3484c758724750d4a737197d0523f9f0f1758b