devtools/jiri-test: add a dummy test to testdata/generate/transitive/lib
This is to workaround the "no test files" check at:
https://cs.corp.google.com/#vanadium/release/go/src/v.io/x/devtools/jiri-test/internal/test/go.go&l=667

The check is meant to ignore tests that have "no test files" in their output.
But in this case, the tests (testdata/generate/...) are run inside
another test (jiri-test). We don't want the output of the testdata tests
affect the test result of the outer test, e.g. if "no test files" is found
in the output of testdata/generate/... tests, we don't want to ignore the
whole outer test.

Change-Id: Ie46a6f1cbfa3e0d9733a22201481320e4d196f3c