v.io/devtools/v23-profile/manager: make profile management 'pluggable'.
** third pass, this now works for go and android, including cross compilation **

The profiles library is now in v.io/jiri/lib/profiles.

There is now a 'driver' programme that implements install, etc, but
named x<verb> instead. Cross compilation support is there in outline
and being tested for a new go profile. The design of this is motivated
by my experience with cross-compiling syncbase with cgo. My goal is
to end up with a fully functioning cross compilation environment with
support for cgo.

MultiPart: 2/2

Change-Id: If0155e373e82cd872b86175db2d6da6650e36df1