devtools: Pull out parts of jiri-go into a standalone package.
The motivation for this is to pull out the functionality that will
be required to implement a separate "jiri-dockergo" command that
will use docker images instead of local profiles to build binaries.

More details on that proposal in:
https://docs.google.com/document/d/1Ud-QUVOjsaya57kgq0j24wDwTzKKE7o_PShQQs0DR5w/edit?usp=sharing

After some debate, it was decided that these docker-based builds should
be placed in a separate "jiri-dockergo" command instead of being
triggered by a flag in "jiri-go" (like jiri-go --docker=...). Thus,
common functionality between the two commands was pulled out into a
separate package.

Change-Id: I3bfdcdc35dcc99b02210c6bbea207e894b37a057