devtools/jiri-go|jiri-dockergo: add --extra-ldflags to merge w/ buildmd ldflags
[Based on discussion w/ toddw]

The traditional way to specify ldflags on the go command-line (jiri go
install --ldflags=... <target>) overrides any ldflags settings done in
the jiri tools (and as a consequence, build info metadata is no longer
added to the binary).  The new flag --extra-ldflags is meant to be
merged with the tooling-generated ldflags, instead of replacing them.

Change-Id: Ib2ed557ee62fa6b36f475cceddb1e77a1fdacfe7