v.io/x/devtools: fix nacl profile.
The nacl profile shows up an inconsistency with how
we expect GO environment variables to be used to
control cross compilation. We use GOARCH, GOOS
in jenkins to request 386 builds etc. We also
expect go toolchains for cross-compiled tools
like nacl to be invoked with appropriate values
for GOARCH (amd64p32-nacl). This means that there
is no single policy that can be used for opting
to use environment variables stored in a profile
vs those supplied on the command line. For now,
we stick with having to set these appropriately
on the command line in all cases, but in future
configuration will be added allowing the user
to request that those in the profile be used instead
in order to avoid the error-prone duplication of
values.

MultiPart: 1/3
Change-Id: Iec97f4be6f233f89e6f27676fdc96b89a94b607e