devtools/jiri-v23-profile/{base,go}: don't use 'master' as a version.
A version string should always refer to a specific and unambiguous
version of the code. master, head, tip don't meet this requirement
since they by definition change over time. For profile implementations
we should always refer to a specific version, so if we need 'tip'
at some point in time, we should create a fixed version for it
as I've done in this CL.

Change-Id: I06c3a00b4a24ec98f3dc5affaf798f46156978ca