jiri-test/internal/test: ensure regression test considers snapshots past 11/10
There is flakiness in the way the old and new proxy get set up, causing
the regression test to flake if using dates before 11/10 (when a fix by
Suharsh went in).  We already updated vanadium-regression-test's jenkins
config, but this change is needed to also fix the regression test that
runs in the presubmit.

Change-Id: Id4e90f5e2de203ebd42025c256265bf19b0f95a9