devtools: Change jiri-related tools to use jiri.X
Currently the jiri tool and most tools under x/devtools use
tool.Context as the main argument to hold program state.  We want
to move to a world where all the jiri-related tools use jiri.X to
capture their program state (e.g. the root directory, the
manifest, etc), and all non-jiri-related tools continue to use
tool.Context.

This is the first change in this direction.  It is the most
intrusive, since we need to change tool.Context to jiri.X in a
huge number of places.  Subsequent changes should be incremental.

The jiri.X struct currently contains a Context along with the
Root directory.  We don't use the Root directory yet; that change
can be performed incrementally.

Most of the non-test code was changed using sed scripts.  The
test code required more manual fixes, since the context is
created and used in a variety of ways.

Also removed the "version" subcommand of many tools.  This was
never working correctly anyways, and would always return
"manual-build", so I've just removed it.

MultiPart: 2/2

Change-Id: I4e7b2516daeb346c63332c0816a528a47199c7e1