mojo profile
The mojo profile can be installed for linux/android with:

  jiri v23-profile install --target=amd64-linux mojo

  jiri v23-profile install --target=arm-android mojo

Once the mojo profile is installed, you can build mojo services with:

  jiri go --profiles=mojo,base --target=amd64-linux build foo

  jiri go --profiles=mojo,base --target=arm-android build foo

In addition to the mojo profile, there is also a profile called
"mojo-dev".  This profile sets the same environment variables as the
mojo profile, but points everything to a full mojo checkout that must be
specified in the --mojo-dev.dir flag when installing.

  jiri v23-profile install --mojo-dev.dir=/path/to/mojo --target=amd64-linux mojo-dev

  jiri v23-profile install --mojo-dev.dir=/path/to/mojo --target=arm-android mojo-dev

Building with the mojo-dev profile is the same as with the mojo profile,
substituting "mojo-dev" for "mojo" in the profiles list.

Change-Id: I1124b193512c7b52a3c52df0d500a76342c26565