RelPaths are always relative to JIRI_ROOT.
RelPaths used to be more general in that they could be made
relative to anything.  The common use case was to make then relative to
JIRI_ROOT.

Now, RelPaths must be relative to JIRI_ROOT.  When Expanding the
RelPath to a full absolute path, a "jirix" must be passed in to
provide the value of JIRI_ROOT.

RelPaths moved from profiles package to jiri package.

Also add jiri project to apiCheckProjects.

MultiPart: 2/2
Change-Id: Ic022446a3788e5c1c7f360e8610beedbdaecaef2