v.io/x/devtools/jiri-test: fix tests so that they run locally.
The tests in jiri-test/internal/test have been failing
locally but not on Jenkins for some time, so I decided to
investigate (broken tests are making it harder and harder
for me to deprecate the 'Run' family of methods and use
the Sequence instead).

This turned up a number of problems:

- test/go_test.go has always been broken, or rather
  is now out of sync with go2xunit behaviour. If it
  ever worked is very unclear to me, but the logic
  is clearly wrong. You could argue that go2xunit
  behaves incorrectly (based on my reading of the
  xunit spec), but I decided to fix the test instead.
- these tests always passed on Jenkins because the
  parsing of their test output was broken. Previously
  devtools/internal/xunit would only ever expect
  a single <testsuite> to be returned whereas these
  tests returned multiple, wrapped in a <testsuites>
  element (note the plural). I've fixed xunit to
  try parsing both ways and to return all of the
  testsuites generated by go2xunit. Note, that this
  behaviour seems to be triggered when test output
  is 'nested' - i.e. a test runs a test, and those
  inner tests may fail. Onto the third problem:
- test/go.go is used both for presubmit tests and by
  test/go_test.go. go_test.go is also run by the
  presubmit tests. Thus leading to 'nested' tests
  and consequently nested test output. go2xunit
  can't understand this and treats all test output
  as coming from the same test. I've added an option
  to suppress test output when tests are run via
  test/go.go from test/go_test.go! This hides the
  test output from the inner tests from go2xunit.

I believe the latter of the bugs above is responsible
for the flakiness we see with 'name mismathc' and
'orphan' test errors from go2xunit.

The only tricky change in this CL is the code in test/go.go
that now handles multiple testsuites. Please review that
carefully since it's likely to be hard to debug. Writing
tests for the tests is currently beyond us since it's just
too hard to create the configurations needed to do so.

Change-Id: Ia0b0ca5c6a6a4bc25d4a4b3c60f9ecd5cf05cb36