Use unique keys and not project names to determine equality.
Jiri used to assume that project names were unique.  Once we implement
importing remote manifests, that assumption becomes hard to enforce.

This CL adds a Key() method to projects which returns a new ProjectKey
type.  The key is currently a combination of the project's name and
remote, although this is easy to change in the future.

The Projects map is now a map[ProjectKey]Project.

There is still a lot of code that tries to lookup projects by their
name.  For instance:
* tools and hooks in the manifest both refer to their projects by name
* some command-line tools take a list of project names as arguments

To support these cases, I added two methods to Projects map:
* Find, which returns all projects matching the key or name,
  and
* FindUnique, which returns a single project matching the key
  or name, or an error if none or many are found.

MultiPart: 2/2
Change-Id: Ia89299541b8df9aea55ab772114e88ab520e0554