add support for cross-compiling to android x86_64
Adds a new base profile with Go with a very recent version supporting
android/x86_64 builds.

The libstdc++.so distributed with the Android NDK only has a handful of
symbols. The full library is shipped as libstdc++.a in usr/lib64, so we
add this file directly to the CGO_LDFLAGS for android/amd64 targets.

Change-Id: Ia89c19d0cf452e5bc3b99408209da346e6cfb438