devtools: Add bootstrap_jiri and local "jiri import"
The idea behind bootstrap_jiri is that it's how to initialize
your jiri root directory.  It sets up the jiri binary and shim
script, and prompts you to add the script directory to your PATH.

The "jiri import" command updates the .jiri_manifest file.
Previously it only handled remote imports.  The new form handles
both remote and local imports, by flipping the order of the
[remote] and [manifest] args.  The -mode flag is also renamed to
-overwrite.

MultiPart: 2/2

Change-Id: Ia95f764bc138a1f792f04b91d3d84497ce6435e5