devtools/vmon: calculate and send aggregated metric data to GCM.
The aggregated data will be sent to a separate timeseries. For example,
for "service-latency" timeseries, its aggregated data is in "service-data-agg"
timeseries.

The new timeseries has a new "aggregation" label with possible values such as
"min", "max", and "avg".

In this CL, I also moved the output statements (test.Pass()) around so that they
appear after the data is sent to GCM.

Change-Id: I3b7a7ca92ace7be49cec145a54a324360f733977