devtools: Replace ReadJiriManifest with LocalProjects
This is a case where we don't actually want to read the manifest,
we really just want to find out where the "third-party" project
lives.

FYI there are two other occurrences of ReadJiriManifest under
devtools, both in the presubmit tool.  Both of those cases
actually need to read the manifest; "presubmit query" uses the
manifest to filter out CLs in projects that aren't listed in the
manifest, while "presubmit test" needs the tools listed in the
manifest to rebuild all tools.

Change-Id: I2e2d2ccef6d1df7190ae1605b75114048d2e62a4