bendroid: Tweaks
(1) Avoid hanging on "adb uninstall" when the "-c" flag is provided (and
no devices are connected to adb).
(2) Forcefully set CGO_ENABLED, since it is required (the generated
main.go requires cgo)
(3) If no GOROOT is set, use the "go" binary from the path
(4) Document the two environment variables that MUST be set when using
bendroid

(2), (3), and (4) don't matter when invoking bendroid via:
  jiri run --target=arm-android bendroid ...
But do help with avoiding a dependence on "jiri run".

Change-Id: Icb761ff29d2ca95e1cc696b7e8121c9fcbc7e9d1