jiri: Replace manifest Labels with SnapshotPath.
The term "label" used to refer to two different but related things:
* The label used to create a snapshot (e.g. "rc" or "go-stable"),
  which is only used to determine the location of the snapshot file.
* The filename of the snapshot being checked out
  (e.g. "snapshot/labels/rc/2016-02-10.04\:18"), which got recorded in
  as the "label" attribute in the JIRI_ROOT/.current_manifest file.

Ignoring this confustion for a second, the label attribute in manifests
isn't as useful in the new jiri world, sinse manifest often import other
manifest, which may themselves have labels.

This CL preserves the first usage of "labels", but replaces the second
usage with "snapshotpath", which is a new attribute only set on Snapshot
manifests (which are required to not have imports).

The snapshotpath is analagous to the second usage of "label" above.  The
snapshot path is set when running "jiri snapshot checkout".

This CL also re-enables the vanadium-release manifest check that was
disabled in v.io/c/19778, replacing "label" with "snapshotpath".

MultiPart: 2/2
Change-Id: Icfa331d39b012e51bd7e57c63948aaa9abddc366