jiri-swift: Set TMPDIR before running swift tests.
The jiri-swift tests call "jiri go build" for the iOS architecture,
which puts iOS build artifacts in TMPDIR (since that's where go puts
build artifacts).

But because we run go test in parallel, we can end up in a situation
where we are building for darwin and iOS at the same time, with the same
TMPDIR, causing collisions.

This CL set TMPDIR to something unique before calling "go build" in the
swift tests.

Change-Id: I08203e57e86b6243ccb0fb0ed2dfdce4cda57334