TBR devtools: Add JIRI_PRESERVE_PATH to disable path munging.
Benj reported that his recent vdl tool changes weren't being
picked up by the presubmit system.  We tracked this down to
Nick's recent change that adds .jiri_root/bin to the front of the
PATH when creating a new jiri.X.

The new behavior is that if JIRI_PRESERVE_PATH is set to a
non-empty value, we disable the path munging.  And we set this
environment variable in the "presubmit test" command, to ensure
we're using the binaries built from the presubmit CLs.

Change-Id: Ia4dc94eadf11bb2dd7dc3f481ba44327ed811093
MultiPart: 2/2