bendroid: Re-implement in a much simpler fashion.
Matt and I realized that shared libraries built for Android by the Go
toolchain are executable. Furthermore, we remembered the -exec flag
to "go test". Which means bendroid can be simplified considerably
and now it is a much smaller program that is a prettier shell script.

The only depdendency is on the 'adb' tool. No gradle or Java compilation
required.

Change-Id: I9a2293deba187cec903100348160ba54c5fd2108