devtools/madb: allow specifying default user for each device.
Each Android device may have multiple users, and 'madb user set' and
other sub-commands can be used to specify which user should be used on
a particular device.

Some of the functions which used to reside in 'name.go' are refactored
and moved to 'madb.go', and used by both 'madb name' and 'madb user'.
Some variable are also renamed to make them more meaningful.

Change-Id: I5cb360a7a4da93d12c6a277055c00cfb8658471f