jiri-profile-v23: Use Go 1.4.3 and not 1.4.2 for bootstrapping.
For targets where the release version of Go 1.6 is not currently
supported (specifically --target=amd64-ios and other *-ios ones),
the Go profile builds go1.4.2 from sources and then builds the
later version of Go from that. This breaks with XCode 7.3 on
OS X because it runs into https://github.com/golang/go/issues/12345

As a quick fix, use 1.4.3 for bootstrapping.
A more appropriate but involved fix might be to just use the release
Go 1.6 as the bootstrapping compiler for the patched Go 1.6.

Change-Id: Iefac907c0fc33a2505ded7e2f8ed18e6d54df4b3