github-mirror should fetch+reset, not pull.
The github-mirror used to pull from googlesource before pushing to
github.  As a result, if the history on googlesource was re-written,
github-mirror would create a merge commit locally, and then push that
merge commit to github.

Instead of pulling, github-mirror should fetch origin, and then reset
--hard to origin/master.  That way the local repo will always exactly
match googlesource, even if the history changes.

Change-Id: Ic4d8b0926e8aec1446ecc3b2427b804e449cd824