Add google sign-in demo that yields default blessings
The swift project has been refactored to build two different
frameworks: Vanadium & Syncbase. The Syncbase will yield simple APIs
for a drop-in framework that masquerades much of the V23 stack.

The demo project has improved to have a picker UI, app icon & more.
It has gained an additional demo that uses the Google SignIn SDK to
obtain an oauth token, and then use that token to receive a default
blessing from dev.v.io. Currently it is using a non-Google app
setting but this can be easily swapped out in the future.

Currently private/public keys are asking V23 to store them in a
special directory under Application Support with
NSFileProtectionCompleteUntilUnlock, but in the future this will
change to keep the private key in the Apple Keychain and have
configurable security setting. Unfortunately, the environmental flags
are being read in at boot in the CGO library so we aren’t able to
pass the correct directory until the credentials flag has already been
created and loaded.

The project is also putting Alamofire and Google Sign In under the
third-party repo, but in the (short) future we’ll move to a Cocoapods
setup so we don’t have to directly check in this code.

MultiPart: 2/3
Change-Id: I93123497b9391535d6932b0c188639d1d411049d