monitoring: Stop monitoring identity/dev.v.io:u/google
This RPC endpoint no longer serves a purpose. We disabled
it's use in  https://vanadium-review.googlesource.com/#/c/21697/
and in general these RPC endpoints that exchange an OAuth token
for a blessing are no longer useful.

The HTTPS endpoint for exchanging an OAuth token for a blessing is
just as secure and functional. So we keep that and I intend to rip
out the code for this RPC service in a separate change.

Change-Id: I3f21b664733b36dc7c4dd2b51d70416605d07576