jiri-swift: Enable tests
We previously disabled tests as jenkins wasn't able to correctly run
them in parallel (things got clobbered between unrelated builds). This
CL is an attempt to help debug the issue and see if we can re-enable
our Swift Testing.

We break up the jiri-swift tests in terms of architecture/project to
deal with timeout limits per-test function as well as disable redundent
and unnecessary tests (amd64 in cgo build + universal build, shared
libs).

Change-Id: I420a93a6b47b06a24b766550fde73560325739f0